chore: extract inline tests to sibling tests.rs files (server crate)
